+++ test/macchiato/test/async/runner.cljs
@@ -1,5 +1,5 @@
 (ns macchiato.test.async.runner
   (:require [doo.runner :refer-macros [doo-tests]]
-            [macchiato.test.async.core]))
+            [macchiato.test.async.futures]))
 
-(doo-tests 'macchiato.test.async.core)
+(doo-tests 'macchiato.test.async.futures)